Move PeriodicStream into periodicsequence.
