+++ sdks/python/apache_beam/transforms/periodicsequence_test.py
@@ -28,9 +28,14 @@ from apache_beam.io.restriction_trackers import OffsetRange
 from apache_beam.testing.test_pipeline import TestPipeline
 from apache_beam.testing.util import assert_that
 from apache_beam.testing.util import equal_to
+from apache_beam.transforms import trigger
+from apache_beam.transforms import window
 from apache_beam.transforms.periodicsequence import PeriodicImpulse
 from apache_beam.transforms.periodicsequence import PeriodicSequence
+from apache_beam.transforms.periodicsequence import PeriodicStream
 from apache_beam.transforms.periodicsequence import _sequence_backlog_bytes
+from apache_beam.transforms.window import FixedWindows
+from apache_beam.utils.timestamp import Timestamp
 
 # Disable frequent lint warning due to pipe operator for chaining transforms.
 # pylint: disable=expression-not-assigned
@@ -131,6 +136,66 @@ class PeriodicSequenceTest(unittest.TestCase):
         _sequence_backlog_bytes(element, 10000, OffsetRange(1002, 1003)), 0)
     self.assertEqual(
         _sequence_backlog_bytes(element, 10100, OffsetRange(1002, 1003)), 8)
+
+
+class PeriodicStreamTest(unittest.TestCase):
+  def test_processing_time(self):
+    with TestPipeline() as p:
+      ret = (
+          p | PeriodicStream([10, 20, 30], interval=2)
+          | beam.Map(lambda _: time.time())
+          | beam.WindowInto(
+              window.GlobalWindows(),
+              trigger=trigger.Repeatedly(trigger.AfterCount(3)),
+              accumulation_mode=trigger.AccumulationMode.DISCARDING,
+          )
+          | beam.GroupBy()
+          | beam.FlatMap(lambda x: [v - min(x[1]) for v in x[1]]))
+      expected = [0, 2, 4]
+      assert_that(ret, equal_to(expected, lambda x, y: abs(x - y) < 0.5))
+
+  def test_interval(self):
+    with TestPipeline() as p:
+      ret = (
+          p | PeriodicStream([1, 2, 3, 4], interval=0.5)
+          | beam.WindowInto(FixedWindows(0.5))
+          | beam.WithKeys(0)
+          | beam.GroupByKey())
+      expected = [(0, [1]), (0, [2]), (0, [3]), (0, [4])]
+      assert_that(ret, equal_to(expected))
+
+  def test_repeat(self):
+    with TestPipeline() as p:
+      ret = (
+          p | PeriodicStream(
+              [1, 2, 3, 4], interval=0.5, max_duration=3, repeat=True)
+          | beam.WindowInto(FixedWindows(0.5))
+          | beam.WithKeys(0)
+          | beam.GroupByKey())
+      expected = [(0, [1]), (0, [2]), (0, [3]), (0, [4]), (0, [1]), (0, [2])]
+      assert_that(ret, equal_to(expected))
+
+  def test_timestamped_value(self):
+    with TestPipeline() as p:
+      ret = (
+          p | PeriodicStream([(Timestamp(1), 1), (Timestamp(3), 2),
+                              (Timestamp(2), 3), (Timestamp(1), 4)],
+                             interval=0.5)
+          | beam.WindowInto(FixedWindows(0.5))
+          | beam.WithKeys(0)
+          | beam.GroupByKey())
+      expected = [(0, [1, 4]), (0, [2]), (0, [3])]
+      assert_that(ret, equal_to(expected))
+
+  def test_stable_output(self):
+    data = [(Timestamp(1), 1), (Timestamp(2), 2), (Timestamp(3), 3),
+            (Timestamp(6), 6), (Timestamp(4), 4), (Timestamp(5), 5),
+            (Timestamp(7), 7), (Timestamp(8), 8), (Timestamp(9), 9),
+            (Timestamp(10), 10)]
+    expected = [1, 2, 3, 4, 5, 6, 7, 8, 9, 10]
+    with TestPipeline() as p:
+      ret = (p | PeriodicStream(data, interval=0.0001))
+      assert_that(ret, equal_to(expected))
 
 
 if __name__ == '__main__':
